refactor(smart-router): Update imports to remove evm sub path and use main router path instead.
