+++ .changeset/fair-pots-look.md
@@ -1,0 +1,6 @@
+---
+'@pancakeswap/universal-router-sdk': patch
+'@pancakeswap/smart-router': patch
+---
+
+Remove smart router evm sub path
