+++ apis/routing/src/provider.ts
@@ -1,5 +1,5 @@
 import { ChainId, getV3Subgraphs } from '@pancakeswap/chains'
-import { OnChainProvider, SubgraphProvider } from '@pancakeswap/smart-router/evm'
+import { OnChainProvider, SubgraphProvider } from '@pancakeswap/smart-router'
 import { createPublicClient, http } from 'viem'
 import { bsc, bscTestnet, goerli, mainnet } from 'viem/chains'
 import { GraphQLClient } from 'graphql-request'
