+++ apps/web/src/views/Swap/V3Swap/containers/ConfirmSwapModal.tsx
@@ -3,7 +3,7 @@ import { Currency, CurrencyAmount, Token, TradeType } from '@pancakeswap/sdk'
 import { memo, useCallback, useMemo } from 'react'
 
 import { useTranslation } from '@pancakeswap/localization'
-import { SmartRouterTrade } from '@pancakeswap/smart-router/evm'
+import { SmartRouterTrade } from '@pancakeswap/smart-router'
 import { WrappedTokenInfo } from '@pancakeswap/token-lists'
 import { Box, BscScanIcon, Flex, InjectedModalProps, Link } from '@pancakeswap/uikit'
 import { formatAmount } from '@pancakeswap/utils/formatFractions'
