+++ apps/web/src/views/Swap/V3Swap/hooks/useCheckSufficient.ts
@@ -3,7 +3,7 @@ import { useAccount } from 'wagmi'
 import { useSwapState } from 'state/swap/hooks'
 import { useCurrency } from 'hooks/Tokens'
 import { Field } from 'state/swap/actions'
-import { SmartRouterTrade } from '@pancakeswap/smart-router/evm'
+import { SmartRouterTrade } from '@pancakeswap/smart-router'
 import { TradeType } from '@pancakeswap/swap-sdk-core'
 import first from 'lodash/first'
 import { useSlippageAdjustedAmounts } from './useSlippageAdjustedAmounts'
