+++ apps/web/src/views/Swap/V3Swap/hooks/useSwapCallArgumentsV2.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Percent, TradeType } from '@pancakeswap/sdk'
-import { SmartRouterTrade } from '@pancakeswap/smart-router/evm'
+import { SmartRouterTrade } from '@pancakeswap/smart-router'
 import {
   PancakeSwapUniversalRouter,
   Permit2Signature,
