+++ apps/web/src/views/Swap/V3Swap/hooks/useWallchain.ts
@@ -1,6 +1,6 @@
 import { ChainId } from '@pancakeswap/chains'
 import { Currency, Token, TradeType } from '@pancakeswap/sdk'
-import { SmartRouterTrade } from '@pancakeswap/smart-router/evm'
+import { SmartRouterTrade } from '@pancakeswap/smart-router'
 import { useUserSlippage } from '@pancakeswap/utils/user'
 import { FeeOptions } from '@pancakeswap/v3-sdk'
 import { captureException } from '@sentry/nextjs'
