+++ packages/universal-router-sdk/src/entities/types.ts
@@ -1,6 +1,6 @@
 import { PermitSingle } from '@pancakeswap/permit2-sdk'
 import { BigintIsh } from '@pancakeswap/sdk'
-import { SwapOptions } from '@pancakeswap/smart-router/evm'
+import { SwapOptions } from '@pancakeswap/smart-router'
 import { Address } from 'viem'
 
 export interface Permit2Signature extends PermitSingle {
