+++ packages/universal-router-sdk/test/trade.test.ts
@@ -9,7 +9,7 @@ import {
   Route as V2Route,
   Trade as V2Trade,
 } from '@pancakeswap/sdk'
-import { PoolType, SmartRouter, SmartRouterTrade, V2Pool, V3Pool } from '@pancakeswap/smart-router/evm'
+import { PoolType, SmartRouter, SmartRouterTrade, V2Pool, V3Pool } from '@pancakeswap/smart-router'
 import { Pool, Route as V3Route, Trade as V3Trade } from '@pancakeswap/v3-sdk'
 import { Address, WalletClient, isHex, parseEther, parseUnits, stringify } from 'viem'
 import { beforeEach, describe, expect, it } from 'vitest'
